# docs/api-documentations/openapi/asset.openapi.json
{
  "components": {
    "schemas": {
...
      "ConsumableBatchDeleteDto": {
        "properties": {
          "ids": {
            "description": "品种 id 列表（1～100 个，不重复）",
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "ConsumableCreateDto": {
        "properties": {
          "categoryId": {
            "description": "分类 id（消耗品分类下的一级子类）",
            "minimum": 1,
            "type": "number"
          },
          "imageOssKey": {
            "description": "图片对象标识（T4-10 上传返回的 OSS key）",
            "maxLength": 200,
            "type": "string"
          },
          "maxHolding": {
            "description": "同时持有上限（借还用品必填，≥1）",
            "minimum": 1,
            "type": "number"
          },
          "name": {
            "description": "品种名称（删除后同名可再建）",
            "maxLength": 50,
            "type": "string"
          },
          "quotaCycle": {
            "description": "申领上限周期（一次性用品必填）",
            "enum": [
              "MONTH",
              "QUARTER",
              "YEAR"
            ],
            "type": "string"
          },
          "quotaLimit": {
            "description": "周期内数量上限（一次性用品必填，≥1）",
            "minimum": 1,
            "type": "number"
          },
          "referencePrice": {
            "description": "参考单价（元，最多两位小数）",
            "example": "10.00",
            "type": "string"
          },
          "remark": {
            "description": "备注",
            "maxLength": 500,
            "type": "string"
          },
          "returnDays": {
            "description": "归还期限（天；借还用品必填，≥1）",
            "minimum": 1,
            "type": "number"
          },
          "safetyStock": {
            "default": 0,
            "description": "安全库存",
            "minimum": 0,
            "type": "number"
          },
          "type": {
            "description": "品种类型",
            "enum": [
              "DISPOSABLE",
              "REUSABLE"
            ],
            "type": "string"
          },
          "unitId": {
            "description": "单位（字典项 id）",
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "type": "object"
      },
      "ConsumableRequestCreateDto": {
        "properties": {
          "idempotencyKey": {
            "description": "客户端为一次用户意图生成的幂等键；网络重试时保持不变",
            "maxLength": 128,
            "type": "string"
          },
          "items": {
            "description": "申领明细行（同一库存条目整单只能一次）",
            "items": {
              "$ref": "#/components/schemas/ConsumableRequestItemDto"
            },
            "maxItems": 100,
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "ConsumableRequestItemDto": {
        "properties": {
          "inventoryItemId": {
            "description": "库存条目 id",
            "minimum": 1,
            "type": "number"
          },
          "purpose": {
            "description": "用途（普通申领必填）",
            "maxLength": 200,
            "type": "string"
          },
          "qty": {
            "description": "数量（正整数）",
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "inventoryItemId",
          "qty",
          "purpose"
        ],
        "type": "object"
      },
      "ConsumableUpdateDto": {
        "properties": {
          "categoryId": {
            "description": "分类 id",
            "minimum": 1,
            "type": "number"
          },
          "imageOssKey": {
            "description": "图片对象标识",
            "maxLength": 200,
            "type": "string"
          },
          "maxHolding": {
            "description": "同时持有上限（借还用品必填）",
            "minimum": 1,
            "type": "number"
          },
          "name": {
            "description": "品种名称",
            "maxLength": 50,
            "type": "string"
          },
          "quotaCycle": {
            "description": "申领上限周期（一次性用品必填）",
            "enum": [
              "MONTH",
              "QUARTER",
              "YEAR"
            ],
            "type": "string"
          },
          "quotaLimit": {
            "description": "周期内数量上限（一次性用品必填）",
            "minimum": 1,
            "type": "number"
          },
          "referencePrice": {
            "description": "参考单价（元，最多两位小数）",
            "example": "10.00",
            "type": "string"
          },
          "remark": {
            "description": "备注",
            "maxLength": 500,
            "type": "string"
          },
          "returnDays": {
            "description": "归还期限（天；借还用品必填）",
            "minimum": 1,
            "type": "number"
          },
          "safetyStock": {
            "description": "安全库存",
            "minimum": 0,
            "type": "number"
          },
          "status": {
            "description": "状态（停用后不可新建入库/申领，既有库存与待审批不受影响）",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ],
            "type": "string"
          },
          "type": {
            "description": "品种类型（只读语义；服务端以既有类型为准）",
            "enum": [
              "DISPOSABLE",
              "REUSABLE"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "safetyStock",
          "status",
          "type"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF 双提交头（状态变更请求必须与 wbme_csrf Cookie 一致）",
        "in": "header",
        "name": "X-WBME-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "服务端会话 Cookie（HttpOnly；登录/激活成功后下发）",
        "in": "cookie",
        "name": "wbme_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "asset 部署单元（资产系统）OpenAPI 契约。构建期生成（主 PRD §9.5）：统一错误结构见 components.schemas.ErrorResponse；默认会话 Cookie + CSRF 头认证，公开路由无 security。",
    "title": "WBME asset API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/consumable-requests": {
      "get": {
        "operationId": "ClaimController_listHistory",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "申请类型",
            "in": "query",
            "name": "requestType",
            "required": false,
            "schema": {
              "enum": [
                "CONSUMABLE_REQUEST",
                "AGENT_REQUEST"
              ],
              "type": "string"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名（范围历史筛选）",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Claim"
        ]
      },
      "post": {
        "operationId": "ClaimController_submit",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableRequestCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Claim"
        ]
      }
    },
    "/api/v1/consumable-requests/mine": {
      "get": {
        "operationId": "ClaimController_listMine",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "申请类型",
            "in": "query",
            "name": "requestType",
            "required": false,
            "schema": {
              "enum": [
                "CONSUMABLE_REQUEST",
                "AGENT_REQUEST"
              ],
              "type": "string"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名（范围历史筛选）",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Claim"
        ]
      }
    },
    "/api/v1/consumables": {
      "get": {
        "operationId": "ConsumableController_list",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "分类 id",
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "品种类型",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "enum": [
                "DISPOSABLE",
                "REUSABLE"
              ],
              "type": "string"
            }
          },
          {
            "description": "状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "ACTIVE",
                "DISABLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "关键字（名称模糊）",
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "description": "仅显示有可用库存的品种（申领目录）",
            "in": "query",
            "name": "hasAvailableStock",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Consumable"
        ]
      },
      "post": {
        "operationId": "ConsumableController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Consumable"
        ]
      }
    },
    "/api/v1/consumables/{id}": {
      "put": {
        "operationId": "ConsumableController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Consumable"
        ]
      }
    },
    "/api/v1/consumables/batch": {
      "delete": {
        "operationId": "ConsumableController_batchDelete",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableBatchDeleteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Consumable"
        ]
      }
    },
...
  },
  "security": [
    {
      "csrfHeader": [],
      "sessionCookie": []
    }
  ],
  "servers": [],
...
}